Add coverage workflow to measure code quality (#11)
* Add coverage workflow to measure code quality

* Add GRCov configuration and upload coverage reports to Codecov

* add caching to codecov workflow

* check for coverage files

* modified grcov settings

* modified grcov settings

* Update output-path in grcov.yml and codecov.yml for coverage files

* Commit changes to Codecov workflow to use grcov output

* modified grcov settings

* modified grcov settings

* modified grcov settings

* modified grcov settings

* modified grcov settings

* modified grcov settings

* modified grcov settings

* modified grcov settings

* Update codecov.yml to specify async-event-emitter test package and args

* Update grcov and codecov configuration and test dependencies in lib.rs

* Update output-type and remove llvm flag

* Remove unused `Context` import and serialize value directly in emit function

---------
