+++ .github/grcov.yml
@@ -1,0 +1,5 @@
+branch: true
+ignore-not-existing: true
+
+output-type: lcov
+output-path: coverage
