+++ .github/workflows/codecov.yml
@@ -1,0 +1,54 @@
+name: Code Coverage
+on:
+     push:
+          branches:
+               - main
+     pull_request:
+          branches:
+               - main
+
+jobs:
+     lint:
+          runs-on: ubuntu-latest
+          steps:
+               - uses: actions/checkout@v1
+               - uses: actions-rs/toolchain@v1
+                 with:
+                      toolchain: nightly
+                      override: true
+                      components: llvm-tools-preview
+               - uses: actions/cache@v3
+                 with:
+                      path: |
+                           ~/.cargo/bin/
+                           ~/.cargo/registry/index/
+                           ~/.cargo/registry/cache/
+                           ~/.cargo/git/db/
+                           target/
+                      key: ${{ runner.os }}-cargo-${{ hashFiles('**/Cargo.lock') }}-coverage
+
+               - uses: actions-rs/cargo@v1
+
+                 with:
+                      command: test
+                      args: --all-features --no-fail-fast
+                 env:
+                      CARGO_INCREMENTAL: "0"
+                      RUSTFLAGS: "-Zprofile -Ccodegen-units=1 -Cinline-threshold=0 -Clink-dead-code -Coverflow-checks=off -Cpanic=abort -Zpanic_abort_tests"
+                      RUSTDOCFLAGS: "-Zprofile -Ccodegen-units=1 -Cinline-threshold=0 -Clink-dead-code -Coverflow-checks=off -Cpanic=abort -Zpanic_abort_tests"
+               - name: Download grcov
+                 run: |
+                      mkdir -p "${HOME}/.local/bin"
+                      curl -sL https://github.com/mozilla/grcov/releases/download/v0.8.10/grcov-x86_64-unknown-linux-gnu.tar.bz2 | tar jxf - -C "${HOME}/.local/bin"
+                      echo "$HOME/.local/bin" >> $GITHUB_PATH
+               - uses: actions-rs/grcov@v0.1
+                 id: coverage
+                 with:
+                      config: .github/grcov.yml
+
+               - name: List coverage files
+                 run: cat ${{steps.coverage.outputs.report}}  >> coverage.Icov
+               - name: Upload coverage reports to Codecov
+                 uses: codecov/codecov-action@v3
+                 env:
+                      CODECOV_TOKEN: ${{ secrets.CODECOV_TOKEN }}
